Add files via upload
Xylo mass production file produces a series of randomized flat pack modules with a spine that extends off to either 4 or 5 nodes. Each node contains space for magnets so that the modules can self-assemble upon exposure to hot water. When treated with a temperature change, the modules thermally expand and connect into a much larger network. This script makes is easy to 3D print rapidly and produce this emergent process.